+++ apps/ai-dial-admin/src/components/Common/SchemaGrid/SchemaTree.tsx
@@ -5,10 +5,10 @@ import { FC, useCallback, useMemo, useState } from 'react';
 import { DialNoDataContent } from '@epam/ai-dial-ui-kit';
 import { JSONSchema7 } from 'json-schema';
 
-import { schemaToTreeNodes } from '@/src/components/TestSuites/utils/schema';
 import { TestSuitesI18nKey } from '@/src/constants/i18n';
 import { useI18n } from '@/src/locales/client';
 import SchemaTreeItem from './SchemaTreeItem';
+import { schemaToTreeNodes } from './utils';
 
 export interface SchemaTreeSelectResult {
   expression: string;
